requirements.txt
# Core dependencies
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pyyaml>=6.0

# RSS feed parsing
feedparser>=6.0.0
python-dateutil>=2.8.0

# Web scraping and content extraction
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# NLP and entity extraction
spacy>=3.7.0
# Run after install: python -m spacy download en_core_web_sm

# Semantic similarity and clustering
sentence-transformers>=2.2.0
torch>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0

# CLI and monitoring
rich>=13.7.0
click>=8.1.0

# Logging
loguru>=0.7.0

# Development/Optional
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0